Refactor cyclomatic complexity analysis and enhance pre-commit configuration
- Introduced a new `CyclomaticComplexityAnalyzer` class to centralize cyclomatic complexity analysis for Python, JavaScript, TypeScript, and C#.
- Updated `ReplicheckRunner` to utilize the new analyzer, improving the analysis process and results handling.
- Removed legacy complexity analysis functions from `utils.py` to streamline the codebase.
- Enhanced pre-commit configuration to ensure consistent execution of tests and formatting checks.
- Added comprehensive tests for the new functionality, ensuring reliability across all supported languages.
