+++ replicheck/utils.py
@@ -92,177 +92,6 @@ def compute_severity(value, threshold):
         return "None"
 
 
-# --- Cyclomatic Complexity Analysis ---
-
-
-def _read_file_content(file_path: Path, mode="r", encoding="utf-8"):
-    with open(file_path, mode, encoding=encoding) as f:
-        return f.read()
-
-
-def _analyze_python_cyclomatic_complexity(code: str, file_path: Path, threshold: int):
-    from radon.complexity import cc_visit
-
-    results = []
-    for block in cc_visit(code):
-        if getattr(block, "complexity", 0) >= threshold:
-            results.append(
-                {
-                    "name": getattr(block, "name", "<unknown>"),
-                    "complexity": getattr(block, "complexity", 0),
-                    "lineno": getattr(block, "lineno", None),
-                    "endline": getattr(block, "endline", None),
-                    "file": str(file_path),
-                    "threshold": threshold,
-                    "severity": compute_severity(
-                        getattr(block, "complexity", 0), threshold
-                    ),
-                }
-            )
-    return results
-
-
-def analyze_cyclomatic_complexity(file_path: Path, threshold: int = 10):
-    """
-    Analyze cyclomatic complexity of a Python file using radon.
-    """
-    try:
-        code = _read_file_content(file_path)
-        return _analyze_python_cyclomatic_complexity(code, file_path, threshold)
-    except Exception:
-        return []
-
-
-def _run_node_helper(helper_path: Path, file_path: Path):
-    import subprocess
-
-    try:
-        proc = subprocess.run(
-            ["node", str(helper_path), str(file_path)],
-            capture_output=True,
-            text=True,
-            timeout=20,
-        )
-        return proc
-    except Exception:
-
-        class DummyProc:
-            returncode = 1
-            stdout = ""
-
-        return DummyProc()
-
-
-def _parse_js_complexity_output(proc, file_path, threshold):
-    import json
-
-    results = []
-    if getattr(proc, "returncode", 1) != 0:
-        return []
-    try:
-        functions = json.loads(proc.stdout)
-    except Exception:
-        return []
-    for fn in functions:
-        if fn.get("complexity", 0) >= threshold:
-            results.append(
-                {
-                    "name": fn.get("name", "<anonymous>"),
-                    "complexity": fn.get("complexity"),
-                    "lineno": fn.get("lineno"),
-                    "endline": fn.get("endline"),
-                    "file": str(file_path),
-                    "threshold": threshold,
-                    "severity": compute_severity(fn.get("complexity", 0), threshold),
-                }
-            )
-    return results
-
-
-def analyze_js_cyclomatic_complexity(file_path: Path, threshold: int = 10):
-    """
-    Analyze cyclomatic complexity of a JS file using typhonjs-escomplex via Node.js helper script.
-    """
-    from pathlib import Path
-
-    helper_path = Path(__file__).parent.parent / "utils" / "helpers.js"
-    try:
-        proc = _run_node_helper(helper_path, file_path)
-        return _parse_js_complexity_output(proc, file_path, threshold)
-    except Exception:
-        return []
-
-
-def _run_cs_analyzer(exe_path: Path, file_path: Path):
-    import subprocess
-
-    try:
-        proc = subprocess.run(
-            [str(exe_path), str(file_path)],
-            capture_output=True,
-            text=True,
-            timeout=20,
-        )
-        return proc
-    except Exception:
-
-        class DummyProc:
-            returncode = 1
-            stdout = ""
-
-        return DummyProc()
-
-
-def _parse_cs_complexity_output(proc, file_path, threshold):
-    import json
-
-    results = []
-    if getattr(proc, "returncode", 1) != 0:
-        return []
-    try:
-        functions = json.loads(proc.stdout)
-    except Exception:
-        return []
-    for fn in functions:
-        if fn.get("Complexity", 0) >= threshold:
-            results.append(
-                {
-                    "name": fn.get("Name", "<anonymous>"),
-                    "complexity": fn.get("Complexity"),
-                    "lineno": fn.get("LineNo"),
-                    "endline": fn.get("EndLine"),
-                    "file": str(file_path),
-                    "threshold": threshold,
-                    "severity": compute_severity(fn.get("Complexity", 0), threshold),
-                }
-            )
-    return results
-
-
-def analyze_cs_cyclomatic_complexity(file_path: Path, threshold: int = 10):
-    """
-    Analyze cyclomatic complexity of a C# file using the compiled ComplexityAnalyzer.exe.
-    """
-    from pathlib import Path
-
-    exe_path = (
-        Path(__file__).parent.parent
-        / "utils"
-        / "C#"
-        / "ComplexityAnalyzer"
-        / "bin"
-        / "Release"
-        / "net9.0"
-        / "win-x64"
-        / "ComplexityAnalyzer.exe"
-    )
-    try:
-        proc = _run_cs_analyzer(exe_path, file_path)
-        return _parse_cs_complexity_output(proc, file_path, threshold)
-    except Exception:
-        return []
-
-
 # --- Large Files ---
